requirements/dev.txt
Django==1.6.2
Markdown==2.4.1
South==0.8.4
django-filter==0.7
djangorestframework==2.4.2
wsgiref==0.1.2
